refactor: queries & mutations
