+++ src/app/home/page.tsx
@@ -1,8 +1,8 @@
 import { Link } from "react-router-dom";
 import { useWorkspaceStore } from "@/state/stores/workspace-store";
-import { useWorkspaces } from "./workspace-query";
+import { useWorkspaces } from "../../data/workspaces/all-workspaces-query";
 import { routePaths } from "@/routes/route-paths";
-import { useCurrentUser } from "@/data/user/active-user-query";
+import { useCurrentUser } from "@/data/users/active-user-query";
 import Welcome from "@/components/Welcome";
 import Skeleton from "react-loading-skeleton";
 import WorkspaceSkeletonCard from "./WorkspaceSkeletonCard";
