+++ src/data/pipelines/detail-pipelines-query.ts
