+++ src/data/stack-components/detail-components-query.ts
