+++ src/data/workspaces/detail-workspaces-query.ts
